size_t fits here better :)
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@19852 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/thing/object/thing.cpp
@@ -1081,7 +1081,7 @@ void csThingStatic::FillRenderMeshes (
 
     rm->mixmode = mixmode; 
     iMaterialWrapper* material = pg.material;
-    for (int m = 0; m < repMaterials.Length(); m++)
+    for (size_t m = 0; m < repMaterials.Length(); m++)
     {
       if (repMaterials[m].old_mat == material)
       {
